requirements-plone43-python2.txt
AccessControl==3.0.11
Acquisition==2.13.12
Babel==1.3
DateTime==3.0.3
DocumentTemplate==2.13.4
ExtensionClass==2.13.2
Markdown==2.0.3
Missing==2.13.1
MultiMapping==2.13.0
Persistence==2.13.2
Plone==4.3.18
Products.ATContentTypes==2.1.19
Products.ATReferenceBrowserWidget==3.0
Products.Archetypes==1.9.18
Products.BTreeFolder2==2.13.3
Products.CMFActionIcons==2.1.3
Products.CMFCalendar==2.2.3
Products.CMFCore==2.2.10
Products.CMFDefault==2.2.4
Products.CMFDiffTool==2.2.1
Products.CMFDynamicViewFTI==4.1.7
Products.CMFEditions==2.2.23
Products.CMFFormController==3.0.8
Products.CMFPlacefulWorkflow==1.5.15
Products.CMFPlone==4.3.18
Products.CMFQuickInstallerTool==3.0.16
Products.CMFUid==2.2.2
Products.DCWorkflow==2.2.5
Products.DateRecurringIndex==2.1
Products.ExtendedPathIndex==3.2.0
Products.ExternalEditor==1.1.3
Products.ExternalMethod==2.13.1
Products.GenericSetup==1.8.9
Products.MIMETools==2.13.0
Products.MailHost==2.13.2
Products.Marshall==2.1.4
Products.MimetypesRegistry==2.0.11
Products.OFSP==2.13.2
Products.PasswordResetTool==2.0.19
Products.PlacelessTranslationService==2.0.7
Products.PloneLanguageTool==3.2.9
Products.PlonePAS==5.1.0
Products.PluggableAuthService==1.11.0
Products.PluginRegistry==1.4
Products.PortalTransforms==2.1.12
Products.PythonScripts==2.13.2
Products.ResourceRegistries==2.2.13
Products.SecureMailHost==1.1.2
Products.Sessions==3.0
Products.StandardCacheManagers==2.13.1
Products.TemporaryFolder==3.0
Products.TinyMCE==1.3.28
Products.ZCTextIndex==2.13.5
Products.ZCatalog==2.13.29
Products.ZSQLMethods==2.13.5
Products.ZopeVersionControl==1.1.3
Products.contentmigration==2.1.19
Products.statusmessages==4.1.2
Products.validation==2.0.2
Pygments==2.0.2
Record==2.13.0
RestrictedPython==3.6.0
Unidecode==0.4.1
ZConfig==2.9.3
ZODB3==3.10.7
ZServer==3.0
Zope2==2.13.28
ZopeUndo==2.12.0
archetypes.querywidget==1.1.3
archetypes.referencebrowserwidget==2.5.10
archetypes.schemaextender==2.1.8
astroid==1.6.6
backports.functools-lru-cache==1.5
bleach==3.1.0
borg.localrole==3.0.3
certifi==2019.3.9
chardet==3.0.4
check-manifest==0.38
collective.elephantvocabulary==0.2.5
collective.monkeypatcher==1.1.5
collective.recipe.template==1.13
collective.z3cform.datetimewidget==1.2.8
colorama==0.4.1
configparser==3.5.3
coverage==4.2
createcoverage==1.5
diazo==1.1.3
docutils==0.14
enum34==1.1.6
experimental.cssselect==0.3
feedparser==5.0.1
five.customerize==1.1
five.formlib==1.0.4
five.globalrequest==1.0
five.intid==1.0.3
five.localsitemanager==2.0.6
future==0.13.1
futures==3.2.0
i18ndude==4.4.0
icalendar==3.10
idna==2.8
initgroups==2.13.0
isort==4.3.20
lazy-object-proxy==1.4.1
mailinglogger==3.7.0
mccabe==0.6.1
mechanize==0.2.5
mr.developer==1.38
ordereddict==1.1
pkginfo==1.5.0.1
plone.alterego==1.1.1
plone.app.blob==1.5.19
plone.app.caching==1.1.12
plone.app.collection==1.0.15
plone.app.content==2.1.7
plone.app.contentlisting==1.0.7
plone.app.contentmenu==2.0.12
plone.app.contentrules==3.0.9
plone.app.contenttypes==1.1.6
plone.app.controlpanel==2.3.11
plone.app.customerize==1.2.3
plone.app.dexterity==2.0.19
plone.app.discussion==2.2.20
plone.app.event==1.1.12
plone.app.folder==1.1.2
plone.app.form==2.2.7
plone.app.i18n==2.0.3
plone.app.imaging==1.0.13
plone.app.intid==1.0.5
plone.app.iterate==2.1.18
plone.app.jquery==1.7.2.1
plone.app.jquerytools==1.9.4
plone.app.layout==2.3.18
plone.app.linkintegrity==1.5.10
plone.app.locales==4.3.16
plone.app.openid==2.0.4
plone.app.portlets==2.5.6
plone.app.querystring==1.2.12
plone.app.redirector==1.2.2
plone.app.registry==1.2.5
plone.app.relationfield==1.2.3
plone.app.robotframework==1.2.1
plone.app.search==1.1.12
plone.app.testing==4.2.6
plone.app.textfield==1.2.11
plone.app.theming==1.1.8
plone.app.upgrade==1.4.5
plone.app.users==1.2.5
plone.app.uuid==1.1.3
plone.app.versioningbehavior==1.2.10
plone.app.viewletmanager==2.0.11
plone.app.vocabularies==2.1.25
plone.app.workflow==2.1.9
plone.app.z3cform==0.7.8
plone.autoform==1.6.2
plone.batching==1.0.8
plone.behavior==1.2.1
plone.browserlayer==2.1.7
plone.cachepurging==1.0.15
plone.caching==1.0.1
plone.contentrules==2.0.7
plone.dexterity==2.2.8
plone.event==1.3.4
plone.fieldsets==2.0.3
plone.folder==1.0.11
plone.formwidget.autocomplete==1.3.0
plone.formwidget.contenttree==1.1.0
plone.formwidget.datetime==1.3.4
plone.formwidget.namedfile==1.0.15
plone.formwidget.querystring==1.1.10
plone.formwidget.recurrence==1.2.6
plone.i18n==2.0.11
plone.indexer==1.0.4
plone.intelligenttext==2.2.1
plone.keyring==3.0.2
plone.locking==2.0.10
plone.memoize==1.1.2
plone.namedfile==3.0.11
plone.openid==2.0.4
plone.outputfilters==1.15.3
plone.portlet.collection==2.1.10
plone.portlet.static==2.0.4
plone.portlets==2.2.3
plone.protect==2.0.3
plone.recipe.zope2instance==4.4.0
plone.registry==1.0.5
plone.resource==1.2.3
plone.resourceeditor==1.0.1
plone.rest==1.4.0
plone.rfc822==1.1.3
plone.scale==1.4.2
plone.schemaeditor==1.4.1
plone.session==3.5.6
plone.stringinterp==1.0.14
plone.subrequest==1.8.5
plone.supermodel==1.2.7
plone.synchronize==1.0.3
plone.testing==4.1.2
plone.theme==2.1.5
plone.transformchain==1.2.2
plone.uuid==1.0.5
plone.z3cform==0.8.1
plonetheme.classic==1.4
plonetheme.sunburst==1.4.8
pycodestyle==2.5.0
pydocstyle==3.0.0
pyflakes==2.1.1
pylama==7.7.1
pylint==1.9.4
pyroma==2.5
python-dateutil==1.5
python-gettext==1.0
python-openid==2.2.5
pytz==2013b0
readme-renderer==24.0
repoze.xmliter==0.6
requests==2.22.0
requests-toolbelt==0.8.0
robotframework==3.0.4
robotframework-debuglibrary==0.4
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==3.3.1
robotsuite==1.7.0
roman==1.4.0
scimschema==0.2.0.dev0
selenium==3.141.0
setuptools==40.8.0
singledispatch==3.4.0.3
snowballstemmer==1.2.1
tempstorage==2.12.2
toml==0.10.0
tqdm==4.32.1
transaction==1.1.1
twine==1.9.1
unittest2==0.5.1
webencodings==0.5.1
wheel==0.33.4
wicked==1.1.12
wrapt==1.11.1
z3c.autoinclude==0.3.7
z3c.caching==2.0
z3c.form==3.2.11
z3c.formwidget.query==0.16
z3c.objpath==1.1
z3c.relationfield==0.6.3
z3c.zcmlhook==1.0b1
zExceptions==2.13.0
zLOG==2.11.2
zc.lockfile==1.0.2
zc.recipe.testrunner==1.2.1
zc.relation==1.0
zdaemon==2.0.7
zest.pocompile==1.4
zest.releaser==6.13.5
zope.annotation==3.5.0
zope.app.container==3.9.2
zope.app.content==3.5.1
zope.app.file==3.6.1
zope.app.form==4.0.2
zope.app.intid==3.7.1
zope.app.locales==3.7.5
zope.app.publication==3.12.0
zope.authentication==3.7.1
zope.broken==3.6.0
zope.browser==1.3
zope.browsermenu==3.9.1
zope.browserpage==3.12.2
zope.browserresource==3.10.3
zope.cachedescriptors==3.5.1
zope.component==3.9.5
zope.componentvocabulary==1.0.1
zope.configuration==3.7.4
zope.container==3.11.2
zope.contentprovider==3.7.2
zope.contenttype==3.5.5
zope.copy==3.5.0
zope.copypastemove==3.7.0
zope.datetime==3.4.1
zope.deferredimport==3.5.3
zope.deprecation==3.4.1
zope.dottedname==3.4.6
zope.dublincore==3.7.0
zope.error==3.7.4
zope.event==3.5.2
zope.exceptions==3.6.2
zope.filerepresentation==3.6.1
zope.formlib==4.3.0
zope.globalrequest==1.2
zope.i18n==3.7.4
zope.i18nmessageid==3.5.3
zope.interface==3.6.7
zope.intid==3.7.2
zope.keyreference==3.6.4
zope.lifecycleevent==3.6.2
zope.location==3.9.1
zope.pagetemplate==3.6.3
zope.processlifetime==1.0
zope.proxy==3.6.1
zope.ptresource==3.9.0
zope.publisher==3.12.6
zope.ramcache==1.0
zope.schema==4.2.2
zope.security==3.7.4
zope.sendmail==3.7.5
zope.sequencesort==3.4.0
zope.site==3.9.2
zope.size==3.4.1
zope.structuredtext==3.5.1
zope.tal==3.5.2
zope.tales==3.5.3
zope.testbrowser==3.11.1
zope.testing==3.9.7
zope.testrunner==4.4.4
zope.traversing==3.13.2
zope.viewlet==3.7.2
